requirements.txt
python-telegram-bot==20.7
openai==1.3.7
google-generativeai==0.3.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
loguru==0.7.2
psutil==5.9.6
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
# 구글 드라이브 API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
